requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.3
scipy>=1.10
pyarrow>=14.0
torch>=2.1
pyqlib>=0.9.3
